getting rid of all relative paths reorganizing component folder hierarchy
