+++ src/components/AddOperatorsCell.vue
@@ -1,9 +1,9 @@
 <script lang="ts" setup>
 import { PropType, computed } from 'vue';
 import { storeToRefs } from 'pinia';
-import { usePlannerStore } from '../store/planner-store';
-import { Operator } from '../types/outputdata';
-import ImageFinder from './ImageFinder.vue';
+import { usePlannerStore } from '@/store/planner-store';
+import { Operator } from '@/types/outputdata';
+import ImageFinder from '@/components/ImageFinder.vue';
 
 const { operator } = defineProps({
     operator: {
