+++ src/components/MissingItems.vue
@@ -1,12 +1,12 @@
 <script setup lang="ts">
-import { usePlannerStore } from '../store/planner-store';
+import { usePlannerStore } from '@/store/planner-store';
 import { storeToRefs } from 'pinia';
-import ItemsDisplay from './ItemsDisplay.vue';
+import ItemsDisplay from '@/components/ItemsDisplay.vue';
 import { computed, ref, watch } from 'vue';
-import PlannerSection from './PlannerSection.vue';
-import { inventoryToList } from '../store/store-inventory-functions';
-import ReservedItemsModal from './ReservedItemsModal.vue';
-import ImportExportModal from './ImportExportModal.vue';
+import PlannerSection from '@/components/PlannerSection.vue';
+import { inventoryToList } from '@/store/store-inventory-functions';
+import ReservedItemsModal from '@/components/ReservedItemsModal.vue';
+import ImportExportModal from '@/components/ImportExport/ImportExportModal.vue';
 
 const { itemsToCraft, itemsToFarm, totalCosts, neededItems, items } = storeToRefs(usePlannerStore());
