+++ src/components/controls/OperatorControlsTabs.vue
