+++ src/components/controls/OperatorLevelPromotion.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
 import { computed, watch } from 'vue';
-import type { SelectedOperator } from '../../types/planner-types';
-import OperatorPromotion from './OperatorPromotion.vue';
-import OperatorLevel from './OperatorLevel.vue';
+import type { SelectedOperator } from '@/types/planner-types';
+import OperatorPromotion from '@/components/controls/OperatorPromotion.vue';
+import OperatorLevel from '@/components/controls/OperatorLevel.vue';
 
 const props = defineProps<{
     selectedOperator: SelectedOperator;
