+++ src/components/generic/NavTabList.vue
