+++ src/main.ts
@@ -1,6 +1,6 @@
 import { createApp } from 'vue'
-import './style.css'
-import App from './App.vue'
+import '@/style.css'
+import App from '@/App.vue'
 import { createPinia } from 'pinia'
 
 import "bootstrap/dist/css/bootstrap.min.css"
@@ -8,9 +8,9 @@ import "bootstrap"
 
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
 import { library } from '@fortawesome/fontawesome-svg-core'
-import icons from './font-awesome-icons'
+import icons from '@/font-awesome-icons'
 
-import './css/night-mode.css';
+import '@/css/night-mode.css';
 
 library.add(...icons);
